Refactoring a bit some components into little smaller ones
